chore: move alert
